+++ packages/core/src/default-settings.js
@@ -1,0 +1,25 @@
+export const DEFAULT_SETTINGS = {
+  animated: false,
+  persistRandomOnExport: true,
+  optimize: true,
+  hideFeedback: false,
+  hideNavigation: false,
+  hidePresets: false,
+  hideScaleToFit: false,
+  initialScaleToFit: true,
+  hideAutoRefresh: false,
+  initialAutoRefresh: true,
+  hideGenerate: false,
+  showMultipleExports: false,
+  ignoreStyles: false,
+  frameRate: 60,
+  videoFormat: "mp4",
+  mp4Settings: {
+    keyFramePeriod: 20,
+    quantizationParameter: 23,
+    speed: 5
+  },
+  webMSettings: {
+    quality: 0.95
+  }
+};