+++ packages/core/src/mechanic-utils.js
@@ -3,6 +3,9 @@ import { toPng, toCanvas } from "html-to-image";
 import { parse as parseCSS, generate as generateCSS } from "css-tree";
 import { selectOne as cssSelectOne } from "css-select";
 import cssSelectBrowserAdapter from "css-select-browser-adapter";
+import deepmerge from "deepmerge";
+
+import { DEFAULT_SETTINGS } from "./default-settings.js";
 
 /**
  * Checks whether a DOM element is instance of SVGElement
@@ -211,6 +214,19 @@ const getTimeStamp = () => {
 };
 
 /**
+ * Creates a standardized settings object from user-provided settings.
+ *
+ * @param {object} settings - User-provided settings
+ * @returns {object} - Standardized settings object
+ */
+const mergeWithDefaultSettings = (
+  settings,
+  defaultSettings = DEFAULT_SETTINGS
+) => {
+  return deepmerge(defaultSettings, settings);
+};
+
+/**
  * Turns an input string into a hash to be used as a cache key.
  * @param {string} str - String to hash
  * @param {number} length - Length of the hash
@@ -253,6 +269,7 @@ export {
   extractSvgSize,
   dataUrlToCanvas,
   getTimeStamp,
+  mergeWithDefaultSettings,
   hashFromString,
   clamp
 };