+++ packages/engine-canvas/CHANGELOG.md
@@ -9,7 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Expose `getCanvas` util to design function. This will create a preview canvas that is adjusted to render crisp on the current display’s pixelDensity.
+- Expose `drawLoop` util to design function. This will call a callback determined by the `frameRate` and pass `frameCount` and `timestamp` in as arguments.
+
+### Changed
+
 - Expose `isPreview` to design function.
+- Updated engine to new `registerFrameCallback` and `registerDoneCallback` setup
+- `done` and `frame` are passed into the design function directly now. Destructuring them from `mechanic` is deprecated.
 
 ## 2.0.0-beta.0 - 2022-03-16